Remove old Ripple class
