--- web/components/animations/Ripple/Ripple.sass
@@ -1,16 +1,0 @@
-span.ripple
-  position: relative
-  display: inline-block
-  overflow: hidden
-  padding-left: 5px
-  padding-right: 5px
-  padding-bottom: 5px
-
-  > div.pop
-    background-color: $pop
-    position: absolute
-    border-radius: 50%
-    pointer-events: none
-    width: 35px
-    height: 35px
-    z-index: 100000
