--- web/components/animations/index.js
@@ -1,5 +1,0 @@
-import Ripple from './Ripple';
-
-export {
-  Ripple,
-};
