--- web/components/animations/index.sass
@@ -1,1 +1,0 @@
-@import ./Ripple/Ripple
