+++ web/components/input/buttons/index.sass
@@ -36,17 +36,9 @@ button
     outline: none
 
   > .ripple
-	  display: block
+    background-color: $pop
     position: absolute
-	  background: $pop
-	  border-radius: 100%
-	  transform: scale(0)
-
-  &.btn-animate .ripple
-    animation: rippleAnimation 0.65s linear
-
-@keyframes rippleAnimation
-	/*scale the element to 250% to safely cover the entire link and fade it out*/
-	100% 
-    opacity: 0
-    transform: scale(2.5)
+    border-radius: 50%
+    pointer-events: none
+    width: 35px
+    height: 35px
